auth ready

+++ src/rooms/rooms.controller.ts
@@ -10,18 +10,27 @@ import {
   HttpStatus,
   UsePipes,
   ValidationPipe,
+  UseGuards,
+  Options,
 } from '@nestjs/common';
 import { RoomsService } from './rooms.service';
 import { CreateRoomDto } from './dto/create-room.dto';
 import { UpdateRoomDto } from './dto/update-room.dto';
+import { JwtAuthGuard } from 'src/auth/guards/jwt.guard';
+import { UserService } from 'src/user/user.service';
+import { UserRole } from 'src/decorators/user-email.decorator';
+
 
 
 @Controller('rooms')
 export class RoomsController {
-  constructor(private readonly roomsService: RoomsService) {}
+  constructor(private readonly roomsService: RoomsService,
+    private readonly userService: UserService) { }
 
   @UsePipes(new ValidationPipe())
   @Post()
+
+  @UseGuards(JwtAuthGuard)
   async create(@Body() createRoomDto: CreateRoomDto) {
     const { number } = createRoomDto;
     const room = await this.roomsService.findByNumber(number);
@@ -36,31 +45,58 @@ export class RoomsController {
   }
 
 
-  @Get(':id')
+/*   @Get(':id')
   async findOne(@Param('id') _id: string) {
     return await this.roomsService.findById(_id);
-  }
+  } */
 
   @Get()
   async findOneByNumber(@Body() createRoomDto: CreateRoomDto) {
     const { number } = createRoomDto;
-    return this.roomsService.findByNumber(number);
+    const room = await this.roomsService.findByNumber(number);
+    console.log(room)
+    if (!room) {
+      throw new HttpException(
+        `Комната с  №${number} не найдена`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    return room;
   }
-
+  @UseGuards(JwtAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateRoomDto: UpdateRoomDto) {
+  async update(@Param('id') id: string, @Body() updateRoomDto: UpdateRoomDto, @UserRole() email: string) {
+    const role = (await this.userService.findUser(email)).role;
+    if (role != 'ADMIN') {
+      throw new HttpException(
+        `только администарторы могут удалять комнааты`,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     return this.roomsService.update(id, updateRoomDto);
   }
-
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @UserRole() email: string) {
     const removeRoom = await this.roomsService.findById(id);
+    const role = (await this.userService.findUser(email)).role;
+    if (role != 'ADMIN') {
+      throw new HttpException(
+        `только администарторы могут удалять комнааты`,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     if (!removeRoom) {
       throw new HttpException(
         `Комнаты с таким ${id} не существует`,
         HttpStatus.BAD_REQUEST,
       );
     }
-     return this.roomsService.remove(id); 
+    return this.roomsService.remove(id);
+  }
+
+  @Get('stat')
+  async showStatstic(@Body() dto: CreateRoomDto) {
+    return this.roomsService.agregateRooms(dto)
   }
 }
